+++ example/lib/logic/counter_logic.dart
@@ -1,9 +1,8 @@
-
 import 'package:popsicle/popsicle.dart';
 
 class CounterLogic extends PopsicleState<int> {
   CounterLogic() : super(state: 0);
 
-  void increment() => shift(state ++);
-  void decrement() => shift(state --);
+  void increment() => shift(state++);
+  void decrement() => shift(state--);
 }